feat: 형 변환 제거
자바는 int형끼리의  계산결과는 int형으로 return 하므로 int형으로 강제 형 변환을 할 필요가 없다.

+++ 03/week2/src/day0306/hw01/Exercise2.java
@@ -4,7 +4,7 @@ public class Exercise2 {
     public static void main (String[] args){
         int num1 = 35;
         int num2 = 10;
-        int quotient = (int)(num1/num2);
+        int quotient = (num1/num2);
         int remainder = num1%num2;
 
         System.out.printf("%d를 %d 으로 나눈 결과 몫은 %d 이고 나머지는 %d 입니다.", num1,num2,quotient,remainder);
